backend/requirements.txt
# backend/requirements.txt

fastapi
uvicorn
pymongo==4.8.0
motor>=3.5,<4.0.0
passlib[bcrypt]
bcrypt==3.2.0
python-jose[cryptography]
pydantic
email-validator
pydantic-settings
python-multipart
celery[redis]
redis
# Use the official Pinecone v3 SDK package
pinecone-client>=3.0.0,<4.0.0
# Use HTTP transport only; avoid gRPC to prevent ALTS warnings outside GCP
python-dotenv
httpx>=0.24
google-generativeai
# Optional tracing over HTTP (not required unless enabling OTEL export)
# opentelemetry-exporter-otlp-proto-http
eventlet>=0.33
neo4j>=5.14
dateparser
python-dateutil
jsonschema>=4.22,<5
spacy>=3.7
pytest
pytest-asyncio==0.23.8
PyYAML
prometheus-client
bleach>=6.1.0
# Web scraping for real-time search content
newspaper3k>=0.2.9
beautifulsoup4>=4.12.0
lxml>=4.9.0
# Optional: serpapi for real-time web search (alternative to Google Custom Search)
# google-search-results>=2.4.2  # Uncomment if using SerpAPI

# NOTE: Do not store secrets in this file.
# Configure environment variables in backend/.env (not committed) or Docker env_file.

